# backend/docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Backend API
  description: API documentation for the backend service
  version: "1.0.0"
servers:
  - url: http://localhost:8888
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
  schemas:
...
    ResponseObject:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: Success
        data:
          nullable: true
    ResponsePage:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        pagination:
          $ref: "#/components/schemas/Pagination"
    Pagination:
      type: object
      properties:
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 20
        totalElements:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
...
    LocationRequest:
      type: object
      properties:
        zoneCode:
          type: string
        shelfCode:
          type: string
        binCode:
          type: string
        description:
          type: string
    LocationResponse:
      type: object
      properties:
        id:
          type: integer
        zoneCode:
          type: string
        shelfCode:
          type: string
        binCode:
          type: string
        fullCode:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        maxCapacity:
          type: number
          format: decimal
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # =========================== LOCATION MAP ===========================
    LocationMapResponse:
      type: object
      properties:
        zones:
          type: array
          items:
            $ref: "#/components/schemas/ZoneData"
    ZoneData:
      type: object
      properties:
        zoneCode:
          type: string
        shelves:
          type: array
          items:
            $ref: "#/components/schemas/ShelfData"
    ShelfData:
      type: object
      properties:
        shelfCode:
          type: string
        bins:
          type: array
          items:
            $ref: "#/components/schemas/BinData"
    BinData:
      type: object
      properties:
        id:
          type: integer
        binCode:
          type: string
        fullCode:
          type: string
        productCount:
          type: integer
          format: int64

    # =========================== INVENTORY ===========================
...
    InventorySummaryResponse:
      type: object
      properties:
        totalProducts:
          type: integer
          format: int64
        totalUnits:
          type: integer
          format: int64
        totalStockValue:
          type: number
          format: decimal
        lowStockCount:
          type: integer
          format: int64
        outOfStockCount:
          type: integer
          format: int64
    CategoryStockResponse:
      type: object
      properties:
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
        productCount:
          type: integer
          format: int64
        totalUnits:
          type: integer
          format: int64
        totalStockValue:
          type: number
          format: decimal
...
    StockValueResponse:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        productSku:
          type: string
          nullable: true
        categoryName:
          type: string
          nullable: true
        quantity:
          type: integer
          format: int64
        unitPrice:
          type: number
          format: decimal
        totalValue:
          type: number
          format: decimal
    ActivityResponse:
      type: object
      properties:
        type:
          type: string
        receiptCode:
          type: string
        date:
          type: string
          format: date-time
        counterpartyName:
          type: string
        lineItems:
          type: integer
        totalAmount:
          type: number
          format: decimal
    DeadStockResponse:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        productSku:
          type: string
        serialNumber:
          type: string
        importedAt:
          type: string
          format: date-time
        daysInStock:
          type: integer
          format: int64
        costPrice:
          type: number
          format: decimal

    # =========================== UPLOAD ===========================
...
  parameters:
    pageParam:
      name: page
      in: query
      description: Page number (0-indexed)
      schema:
        type: integer
        default: 0
    sizeParam:
      name: size
      in: query
      description: Page size
      schema:
        type: integer
        default: 20
...
paths:
...
  /api/v1/location:
    post:
      tags: [Location]
      summary: Create a new location
      operationId: createLocation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/LocationResponse"

    get:
      tags: [Location]
      summary: List locations (paginated)
      operationId: listLocations
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: isActive
          in: query
          schema:
            type: boolean
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ResponsePage"
  /api/v1/location/search:
    get:
      tags: [Location]
      summary: Search locations by keyword
      operationId: searchLocations
      security:
        - BearerAuth: []
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ResponsePage"
  /api/v1/location/{id}:
    get:
      tags: [Location]
      summary: Get location by ID
      operationId: getLocation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/LocationResponse"

    put:
      tags: [Location]
      summary: Update location
      operationId: updateLocation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseObject"
  /api/v1/location/{id}/toggle-active:
    put:
      tags: [Location]
      summary: Toggle location active status
      operationId: toggleLocationActive
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseObject"
  /api/v1/location/map:
    get:
      tags: [Location]
      summary: Get zone → shelf → bin tree with product counts
      operationId: getLocationMap
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/LocationMapResponse"

  # =========================== DASHBOARD ===========================
...
  /api/v1/inventory:
    get:
      tags: [Inventory]
      summary: List inventory items (paginated)
      operationId: listInventoryItems
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ResponsePage"

  # =========================== REPORT ===========================
  /api/v1/report/inventory-summary:
    get:
      tags: [Report]
      summary: Get inventory summary
      operationId: getInventorySummary
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/InventorySummaryResponse"
  /api/v1/report/inventory-by-category:
    get:
      tags: [Report]
      summary: Get stock by category
      operationId: getStockByCategory
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/CategoryStockResponse"
  /api/v1/report/low-stock:
    get:
      tags: [Report]
      summary: Get low stock products (paginated)
      operationId: getLowStock
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ResponsePage"
  /api/v1/report/stock-value:
    get:
      tags: [Report]
      summary: Get stock value by product
      operationId: getStockValue
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/StockValueResponse"
  /api/v1/report/activity:
    get:
      tags: [Report]
      summary: Get activity report by date range
      operationId: getActivity
      security:
        - BearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ActivityResponse"
  /api/v1/report/dead-stock:
    get:
      tags: [Report]
      summary: Get dead stock report
      operationId: getDeadStock
      security:
        - BearerAuth: []
      parameters:
        - name: daysThreshold
          in: query
          schema:
            type: integer
            default: 90
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DeadStockResponse"

  # =========================== UPLOAD ===========================
...